+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `play_cli` function
 - `SECURITY.md`
 ### Changed
+- `main` function updated
 - OSs local checklist added to pull request template
 - `README.md` modified
 ## [0.5] - 2024-04-03